requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
httpx
pydantic-settings
python-dotenv
python-jose
pytz
